+++ src/main/java/org/jabref/model/strings/StringUtil.java
@@ -735,4 +735,8 @@ public class StringUtil {
     public static String substringBetween(String str, String open, String close) {
         return StringUtils.substringBetween(str, open, close);
     }
+
+    public static String ignoreCurlyBraces(String title){
+        return isNotBlank(title) ? title.replace("{","").replace("}","") : title;
+    }
 }